refactor package structure for easier management
