+++ src/main/java/com/github/flo456123/BackBond/entry/PYCRRepository.java
@@ -1,13 +1,12 @@
-package com.github.flo456123.BackBond.data.repository;
+package com.github.flo456123.BackBond.entry;
 
-import com.github.flo456123.BackBond.data.model.Entry;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface DailyTreasuryRatesRepository
+public interface PYCRRepository
         extends JpaRepository<Entry, Integer> {
 
     @Query("SELECT e FROM Entry e WHERE e.newDate BETWEEN ?1 AND ?2")
